requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==7.1.2
flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
jinja2==2.11.3
markupsafe==1.1.1
pyyaml==5.4.1
werkzeug==1.0.1
